Fix underflow in Neutronium encoder
Summary:
If the field is not fixed, -1 underflows because the ternary expression gets its
type from the uint32_t field.fixedStringSize

Not sure if the preferable fix is to change all those literals to ##-1L##

Test Plan: no longer saw underflow when using Neutronium, see D1190014

Reviewed By: tudorb@fb.com

FB internal diff: D1190024

+++ thrift/lib/cpp/protocol/neutronium/Schema.cpp
@@ -146,7 +146,7 @@ int64_t Schema::fixedSizeForField(const StructField& field) const {
     case reflection::TYPE_DOUBLE:
       return field.isFixed ? 8 : -1;
     case reflection::TYPE_STRING:
-      return field.isFixed ? field.fixedStringSize : -1;
+      return field.isFixed ? static_cast<int64_t>(field.fixedStringSize) : -1;
     default:
       break;
     }
